Adding function helper file to cards rotuer, so update payload is validated properly
